File names changed and style updated Feature/43.user-page
