--- src/constants/Layout/UserContainer.ts
@@ -1,18 +1,0 @@
-import styled from 'styled-components'
-
-export const UserContainer = styled.div`
-   display: flex;
-   flex-direction: column;
-   align-items: center;
-   justify-content: center;
-   max-width: 60%;
-   margin: 0 auto;
-   border: 1px solid #292a2b;
-   background-color: #222224;
-   color: #f7f7f7;
-`
-export const UserAvatar = styled.img`
-   width: 250px;
-   height: 250px;
-   border-radius: 50%;
-`
